refactor and take all functions into core-functions module
